requirements.txt
# ============================================
# Core Testing Framework
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1

# ============================================
# Code Coverage & Quality
# ============================================
pytest-cov==4.1.0
coverage[toml]==7.3.2

# ============================================
# Parallel Test Execution
# ============================================
pytest-xdist==3.3.1

# ============================================
# Test Reporting
# ============================================
pytest-html==4.0.0
pytest-metadata==3.0.0
allure-pytest==2.13.2

# ============================================
# Mocking & Fixtures Enhancement
# ============================================
pytest-mock==3.11.1
factory-boy==3.3.0

# ============================================
# Test Data Generation
# ============================================
faker==19.6.2
hypothesis==6.87.1

# ============================================
# Parameterization
# ============================================
parameterized==0.9.0

# ============================================
# API Testing (Optional)
# ============================================
requests==2.31.0
httpx==0.25.0

# ============================================
# Logging & Debugging
# ============================================
pytest-icdiff==0.8
pytest-sugar==0.9.7

# ============================================
# Environment & Configuration
# ============================================
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# ============================================
# Code Linting & Type Checking (Dev)
# ============================================
flake8==6.1.0
black==23.9.1
isort==5.12.0
mypy==1.6.1
pylint==3.0.1
